Bug 768950 - TIFF private tags should not generate warnings.
This is a regression from bug 131975. Any unknown tag over 32768 is not
an error. This is a reserved zone where it is allowed to create custom
tags.
The warning indeed changed since libtiff 4.0.0alpha where it has become:
"Unknown field with tag %d (0x%x) encountered"
This explains why it was not recognized anymore.


Former-commit-id: 3cb70e67cce56960ab1055e6ef7f72eff1abf55c