Updated French translation.
"dépôt" is not a proper translation for "commit".
"Modification" seems more appropriate, though the Pro Git book (which
is a reference on the topic) keeps the English word "commit" even in
the French version (simply italicized when in a sentence). But let's
use a French word for now.


Former-commit-id: 2340f37ef4e57e9b9d2a6937adce2cf2d32f63c4