po: explain why `make update-po` will fail on a VPATH build.
Commit c96006919 worked around a still-existing bug (well at least the
bug report is still open) of intltool with automake 1.15 when the build
and source directories are different (VPATH build) by replacing
`intltool-update` by `true`. But because of this `make update-po` fails
on such a build and the make output is totally unclear at why. Add an
error message when this happens.


Former-commit-id: c4bb7089d1b81c26b3890894093828b021f80263