Revert "app: handle gtk_accel_map_change_entry () failure..."
This reverts commit a250c60e25f5df9703dacb8e6425ef667c3fb9f0 [formerly eab4929a782ef63a06284515f8e518592152ee47].

Oups it would seem gtk_accel_map_change_entry() could return FALSE even
when the expected shortcut is correctly set (my guess is that it was
already the same shortcut, so indeed no "change" happened, though it is
not a failure either; yet I haven't checked if that is the actual
reason).
Let's just revert this. It's not always a good thing to be too thorough!
Sorry for this!


Former-commit-id: 4a9a0723f5a4e8cbd5e5a5e5978d01df9bad70a0