imagemap: clear preview area background before drawing image
Previously, we neither cleared the preview area's buffer, nor set
its max size, so if the preview area was bigger than the image, the
rest of the widget would display garbage.

This is an ad-hoc-ish fix, simply filling the preview area's buffer
with the background color prior to drawing the image.  Anything
more sophisticated will have to wait till image maps are cool
again, I guess :)


Former-commit-id: f91e2a94a7d0ab41d0c53a8eae480c35b1c11cf7