build: fix Webkit build in Flatpak package.
For some reason, Webkit would not build with Python 2 on a different
machine though it did on mine. Let's just force Python 3 usage (for
build only, not as a runtime dependency) by setting the $PYTHON
environment variable during this build. Anyway Python 3 is available
from the base build so it's not a new dependency.


Former-commit-id: afc69f83ae160000f73911af056d60b733aa6a63