Revert "Bug 780270 - MinGW build fail on Windows, wrong definition in..."
This reverts commit 2f7007de268ba52ff353bd5bfeca334c25e231d1 [formerly 2226fb1dc814d015977ea6f7d5bedcb6ca557627].
It doesn't make much difference right now, will work the same on
native builds and be broken on cross-builds.
I need to fix how invert-svg is built (broken on commit 84439a8),
so that it is a build executable again. Work in progress.
But since it makes no difference right now, let's just revert this
one already.


Former-commit-id: 228929957af7caaa0fa2fe28f8a8518b881d063e