tools: invert-svg is to be built natively for the build OS.
Partially reverts commit 83ec3aa61a8124d479a1980c84888bd6bd58c6cd [formerly 84439a874860c517838c4d5969b353ced1c31b1a].
The fact that tools/invert-svg was kind of manually compiled was on
purpose. Using the automake syntax with any of the *_PROGRAMS compiles
for the host OS, but this tool is only for compile-time.
This fixes cross-compilation.

As for compute_svg_viewbox, let's not build it for now because its
purpose was to help proper extraction of SVG icons from a single SVG
file. Unfortunately librsvg was not providing acceptable results anyway
so right now all SVG icons are manually exported. We keep the code for
future use when the library will improve.


Former-commit-id: 9a2da538022896565f3a27a66eec1416811b6bb3