autogen: fix and improve automake/aclocal detection.
Forgot to edit the first test in commit 6cf2641. If automake/aclocal
1.11 was installed, autogen.sh breaks, even though other higher versions
are installed too, because it is the first test.
Actually I'm not sure I understand this first test. Is it like the
"preferred" automake version?
Also add a test for generic `automake` binary in order to handle newer
versions. Thanks to Éric Hoffman for this proposition.


Former-commit-id: 8df8cd03f337b67e5547c402e1983d23c8f56e57