configure: remind to run gdk-pixbuf-query-loaders.exe on Windows builds.
Even though the SVG loader is installed, it needs to be properly
registered for GdkPixbuf to find it. This is a problem for Windows where
the installation prefix can end up being anything and where the command
`gdk-pixbuf-query-loaders` is not run by any common component anyway.
Let's just warn the Windows packager to not forget to have the installer
run it if vector icons are enabled.


Former-commit-id: e83b1361081503315fde56ca69dbd3199385d343